requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
requests==2.31.0
transformers==4.35.2
torch==2.1.1
nltk==3.8.1
python-multipart==0.0.6
pydantic==2.5.2
newsapi-python==0.2.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4 